// src/api/response.type.ts
export type AnnouncementItem = {
  title: string
  content: string
  image_path: string
}

export interface Announcement {
  id: number
  type: ANNOUNCEMENT_MEMBER_TYPE.Enums
  start_time: string
  end_time: string
  detail: Record<string, AnnouncementItem>
  display_options: ANNOUNCEMENT_DISPLAY_TYPE.Enums[]
  langDetail?: AnnouncementItem
}

export type AnnouncementList = Announcement[]